added the owner field so that the user can only see his contacts

+++ controllers/contactsControllers.js
@@ -3,7 +3,8 @@ const { HttpError, ctrlWrapper } = require("../helpers");
 const { Contact, addSchema } = require("../model/contactsModel");
 
 const getAllContacts = async (req, res, next) => {
-  const result = await Contact.find();
+  const { _id: owner } = req.user;
+  const result = await Contact.find({ owner });
 
   if (!result.length) {
     throw HttpError(404, "No contacts found");
@@ -13,9 +14,13 @@ const getAllContacts = async (req, res, next) => {
 };
 
 const getContactById = async (req, res, next) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
 
-  const result = await Contact.findById(contactId);
+  const result = await Contact.findById(contactId).populate({
+    path: "owner",
+    match: { _id: owner },
+  });
 
   if (!result) {
     throw HttpError(404, "No contacts found");
@@ -25,9 +30,10 @@ const getContactById = async (req, res, next) => {
 };
 
 const deleteContact = async (req, res, next) => {
+  const { _id: owner } = req.user;
   const { contactId } = req.params;
 
-  const result = await Contact.findOneAndDelete({ _id: contactId });
+  const result = await Contact.findOneAndDelete({ _id: contactId, owner });
 
   if (!result) {
     throw HttpError(404, "No contacts found");
@@ -37,12 +43,13 @@ const deleteContact = async (req, res, next) => {
 };
 
 const createContact = async (req, res, next) => {
+  const { _id: owner } = req.user;
   const { error } = addSchema.validate(req.body);
   if (error) {
     throw HttpError(400, "Missing required field");
   }
 
-  const result = await Contact.create(req.body);
+  const result = await Contact.create({ ...req.body, owner });
 
   res.status(201).json(result);
 };
